3.5.2
Bug Fixes:

Several bugs have been fixed:

Fixed an error that occurred when logging in for the first time and a saved location did not exist

Resolved an issue that prevented the /spawn world command from working

Fixed a bug that caused world aliases to not be displayed in the /spawn world command

Fixed a bug that prevented another player from being teleported to spawn

Deleting Message Files:

It is recommended to delete the message files to avoid possible conflicts and errors in the new version of the plugin. Restart the server after deleting the files to apply the changes correctly.